requirements.txt
olefile==0.44
Pillow==4.0.0
pytesseract==0.1.6
